// src/api/planejador/models/MudancasResDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export enum MudancasResDtoMudancaEnum {
  ATUALIZACAO = 'ATUALIZACAO',
  REMOÇÃO = 'REMO\u00C7\u00C3O',
  INSERÇÃO = 'INSER\u00C7\u00C3O',
}

export type MudancasResDto = {
  /**
   * @type string
   */
  mudanca: MudancasResDtoMudancaEnum;
  /**
   * @type string
   */
  entidade: string;
  /**
   * @type string
   */
  valorAntigo: string;
  /**
   * @type string
   */
  valorNovo: string;
};